T: move all naming inspections into separate package
Also create separate file for each naming inspection test class.
It allows running single tests with gradle
